requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
packaging==23.1
pymongo==3.13.0
pymongo-stubs==0.2.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
